+++ src/main/java/com/craftix/hostile_humans/entity/entities/ModEntityType.java
@@ -28,6 +28,18 @@ public class ModEntityType {
     public static final RegistryObject<EntityType<Human>> ROAMER = ENTITIES.register("human_roamer",
             () -> EntityType.Builder.<Human>of((entityEntityType, level) -> new Human(entityEntityType, level, HumanTier.ROAMER), HumanEntity.CATEGORY)
                     .sized(0.6F, 1.8F).clientTrackingRange(16).build("human_roamer"));
+    public static final RegistryObject<EntityType<Ronin>> RONIN = ENTITIES.register("human_ronin",
+            () -> EntityType.Builder.<Ronin>of(Ronin::new, HumanEntity.CATEGORY)
+                    .sized(0.6F, 1.8F).clientTrackingRange(16).build("human_ronin"));
+    public static final RegistryObject<EntityType<Samurai>> SAMURAI1 = ENTITIES.register("human_samurai1",
+            () -> EntityType.Builder.<Samurai>of((entityType, level) -> new Samurai(entityType, level, HumanTier.SAMURAI1), HumanEntity.CATEGORY)
+                    .sized(0.6F, 1.8F).clientTrackingRange(16).build("human_samurai1"));
+    public static final RegistryObject<EntityType<Samurai>> SAMURAI2 = ENTITIES.register("human_samurai2",
+            () -> EntityType.Builder.<Samurai>of((entityType, level) -> new Samurai(entityType, level, HumanTier.SAMURAI2), HumanEntity.CATEGORY)
+                    .sized(0.6F, 1.8F).clientTrackingRange(16).build("human_samurai2"));
+    public static final RegistryObject<EntityType<Bandit>> BANDIT = ENTITIES.register("human_bandit",
+            () -> EntityType.Builder.<Bandit>of(Bandit::new, HumanEntity.CATEGORY)
+                    .sized(0.6F, 1.8F).clientTrackingRange(16).build("human_bandit"));
     public static final RegistryObject<EntityType<SpawnerEntity>> SPAWNER_ENTITY = ENTITIES.register("human_group",
             () -> EntityType.Builder.of(SpawnerEntity::new, MobCategory.CREATURE)
                     .sized(0.6F, 1.8F).clientTrackingRange(16).build("human_group"));
@@ -41,6 +53,10 @@ public class ModEntityType {
         event.put(HUMAN1.get(), Human.createAttributes().build());
         event.put(HUMAN2.get(), Human.createAttributes().build());
         event.put(ROAMER.get(), Human.createAttributes().build());
+        event.put(RONIN.get(), Human.createAttributes().build());
+        event.put(SAMURAI1.get(), Human.createAttributes().build());
+        event.put(SAMURAI2.get(), Human.createAttributes().build());
+        event.put(BANDIT.get(), Human.createAttributes().build());
         event.put(SPAWNER_ENTITY.get(), Human.createAttributes().build());
 
         if (ModList.get().isLoaded("travelersbackpack")) {