+++ src/main/java/com/craftix/hostile_humans/item/ModItems.java
@@ -26,4 +26,20 @@ public class ModItems {
             () -> new ForgeSpawnEggItem(ModEntityType.ROAMER, MapColor.COLOR_ORANGE.col,
                     MapColor.COLOR_BLACK.col,
                     new Item.Properties().rarity(Rarity.EPIC)));
+    public static final RegistryObject<Item> RONIN_SPAWN_EGG = ITEMS.register("human_ronin_spawn_egg",
+            () -> new ForgeSpawnEggItem(ModEntityType.RONIN, MapColor.COLOR_RED.col,
+                    MapColor.COLOR_BLACK.col,
+                    new Item.Properties().rarity(Rarity.EPIC)));
+    public static final RegistryObject<Item> SAMURAI1_SPAWN_EGG = ITEMS.register("human_samurai1_spawn_egg",
+            () -> new ForgeSpawnEggItem(ModEntityType.SAMURAI1, MapColor.COLOR_RED.col,
+                    MapColor.TERRACOTTA_WHITE.col,
+                    new Item.Properties().rarity(Rarity.EPIC)));
+    public static final RegistryObject<Item> SAMURAI2_SPAWN_EGG = ITEMS.register("human_samurai2_spawn_egg",
+            () -> new ForgeSpawnEggItem(ModEntityType.SAMURAI2, MapColor.COLOR_RED.col,
+                    MapColor.GOLD.col,
+                    new Item.Properties().rarity(Rarity.EPIC)));
+    public static final RegistryObject<Item> BANDIT_SPAWN_EGG = ITEMS.register("human_bandit_spawn_egg",
+            () -> new ForgeSpawnEggItem(ModEntityType.BANDIT, MapColor.COLOR_BLACK.col,
+                    MapColor.GOLD.col,
+                    new Item.Properties().rarity(Rarity.EPIC)));
 }